// src/app/interfaces/productionrecords.ts
export interface MixedSolution {
    id: number;
    productionRecordId: number;
    workerId: number;
    tankNo: string;
    mixingDate: Date;
    solidContext: number;
    referToDDNo: string;
    startUseTime: Date;
    mixedIn: number;
    mixedOut: number;
    vesselWeigth: number;
    referToCNNo: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
}